social_backend/requirements.txt
asgiref==3.11.0
Django==3.2.23
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.8.0
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pymongo==3.12.3
python-decouple==3.8
pytz==2025.2
sqlparse==0.2.4
